Get mission_year from timestamp of panorama Part of DP-6022
The mission year was read from missions.csv and for missing missions was
set to the arbitrary value of 2015. This didn't matter because for
filtering the date in timestamp was used. But when adding mission_year
to panorama it became a way to filter. This change sets the mission_year
to the year of the timestamp of the panorama, even if the mission is not
known.