Reorder migrations to accomodate materialized views
Refreshing materialized views defined by a `select * from` from an
underlying table that has a new attribute doesn't add that attribute to
the refreshed materialized view.

This solves this by reordering migrations in such a way that the field
is added before the materialized views are defined.

This is possible in our set-up, where database is always built from
scratch.