DS-416 Fix geoselection for panorama.
The most recent panorama in region is select, close to the selected point.
However, when selectin newer panoramas that do not exist we should only consider
panorama's with 'done' status, because we return only panaoramas with 'done'.
Otherwise panorama X_2019 is discarded because a newer panorama X_2020 exists with 'rendered' status,
but that panorama is not returned because it has not the 'done' status
status. This is also consistent with the map server layer query